refactor: 모듈별 폴더 생성
